Restructure inspectors and clean up documentation
